+++ history.md
@@ -4,6 +4,7 @@
   from language client option.
 - Add configuration `documentHighlight.timeout`.
 - Add `tabPersist` option to `ListAction`.
+- Add `refactor` to `LocationList`
 
 # 2022-01-30